Redefine all C interface structs
